+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Next
+
+- Updated React Native rules
+
 ## 0.0.3
 
 - Add initial React Native ESLint config